refactor(act): extract runRestore loop helper for all adapters
Adapters were duplicating the iteration loop too — drop_snapshots
filter, on_progress callback, causation-remap via idMap, kept/
dropped counting. All identical across InMemoryStore, PostgresStore,
and SqliteStore. Only the actual INSERT and the transaction
mechanics genuinely differ across backends.

Extracts the iteration to `runRestore(source, opts, writeRow)` in
`libs/act/src/restore-validate.ts`. Adapters provide a `writeRow`
callback that takes a row + already-causation-rewritten meta and
returns the new id; the helper handles everything else. Each
adapter's restore method now looks like:

```ts
const started = Date.now();
// adapter: open transaction + wipe
const partial = await runRestore(source, opts, async (row, meta) => {
  // adapter: insert + return new id
});
// adapter: commit
return { ...partial, duration_ms: Date.now() - started };
```

Net change: -127 / +112 across four files. Per-adapter restore
shrinks from ~50 lines of mixed loop+adapter code to ~20 lines of
pure adapter-specific transaction + write. The shared iteration
contract (when on_progress fires, which rows count as drops, how
causation gets remapped) lives in exactly one place.

`runRestore` is exported publicly from `@rotorsoft/act` so future
third-party adapters get the same iteration contract for free —
implement `writeRow`, get drop/progress/causation behavior
automatically.

## Test plan

- `pnpm typecheck` — clean
- `pnpm test` — 1859 / 1859
- Coverage: 100% statements / 100% branches / 100% functions / 100% lines.
- Function count dropped from 1083 → 1080 (3 duplicate per-adapter
  iteration helpers collapsed into 1 shared helper).

## Stability charter impact

**Charter surface touched: `libs/act/src/index.ts`** indirectly
(re-exports `runRestore` and `RestoreRowWriter` from
`restore-validate.ts`). Additive — new public exports gated by
STABILITY.md's minor-bump policy.
